build: add me_order_book library and wire limit order implementation
Add the static order book target, public headers for Trade and matching, and track order_book.cpp with formatting aligned to the project style.

Made-with: Cursor

+++ include/order.hpp
@@ -10,8 +10,8 @@ enum class Side {
 struct Order {
   uint64_t id;
   Side side;
-  double price;
-  double quantity;
+  uint64_t price;
+  uint64_t quantity;
   std::chrono::system_clock::time_point timestamp;
   bool is_filled;
 };
